Fix issue with outdated query data being shown using ShowLoadingQuery
Also clean up previous commits and fix a couple of small bugs
