+++ src/app/components/pages/GraphSketcher.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useRef, useState} from 'react';
-import {AppState, isaacApi, mutationSucceeded, useAppDispatch, useAppSelector} from "../../state";
+import {AppState, isaacApi, mutationSucceeded, useAppSelector} from "../../state";
 import {Col, Container, Row} from 'reactstrap';
 import {TitleAndBreadcrumb} from '../elements/TitleAndBreadcrumb';
 import {GraphChoiceDTO} from '../../../IsaacApiTypes';
