+++ src/app/components/pages/Topic.tsx
@@ -1,6 +1,6 @@
-import React, {useEffect} from "react"
+import React from "react"
 import {Link, withRouter} from "react-router-dom";
-import {AppState, isaacApi, selectors, useAppDispatch, useAppSelector} from "../../state";
+import {isaacApi, selectors, useAppSelector} from "../../state";
 import {ShowLoading} from "../handlers/ShowLoading";
 import {IsaacContent} from "../content/IsaacContent";
 import {
