+++ src/app/state/index.ts
@@ -38,3 +38,4 @@ export * from "./actions";
 export * from "./selectors";
 export * from "./reducers";
 export * from "./store";
+export * from "./actions/service";
